chore(address): remove duplicate
