+++ next-tavla/app/(admin)/components/TileSelector/utils.ts
@@ -1,6 +1,6 @@
-import { fetchServerQuery } from 'app/(admin)/fetch'
 import { getFormFeedbackForError } from 'app/(admin)/utils'
 import { WalkDistanceQuery } from 'graphql/index'
+import { fetchQuery } from 'graphql/utils'
 import { nanoid } from 'nanoid'
 import { DEFAULT_ORGANIZATION_COLUMNS } from 'types/column'
 import { TLocation } from 'types/meta'
@@ -30,7 +30,7 @@ export async function getWalkingDistance(
 ) {
     if (!stopPlaceId) return getFormFeedbackForError()
     if (!location) return
-    return await fetchServerQuery(WalkDistanceQuery, {
+    return await fetchQuery(WalkDistanceQuery, {
         stopPlaceId: stopPlaceId,
         location: {
             longitude: location.coordinate?.lng ?? 0,
