+++ next-tavla/src/Board/scenarios/QuayTile/index.tsx
@@ -1,12 +1,12 @@
 import { TQuayTile } from 'types/tile'
 import { Table } from '../../scenarios/Table'
 import classes from './styles.module.css'
-import { useQuery } from 'graphql/utils'
 import { GetQuayQuery } from 'graphql/index'
 import { Tile } from 'components/Tile'
 import { TableHeader } from '../Table/components/TableHeader'
 import { isNotNullOrUndefined } from 'utils/typeguards'
 import { TileLoader } from 'Board/components/TileLoader'
+import { useQuery } from 'hooks/useQuery'
 
 export function QuayTile({
     placeId,
