+++ next-tavla/src/Board/scenarios/StopPlaceTile/index.tsx
@@ -1,11 +1,11 @@
 import { TStopPlaceTile } from 'types/tile'
 import { Table } from '../../scenarios/Table'
 import classes from './styles.module.css'
-import { useQuery } from 'graphql/utils'
 import { StopPlaceQuery } from 'graphql/index'
 import { Tile } from 'components/Tile'
 import { TableHeader } from '../Table/components/TableHeader'
 import { TileLoader } from 'Board/components/TileLoader'
+import { useQuery } from 'hooks/useQuery'
 
 export function StopPlaceTile({
     placeId,
