+++ next-tavla/src/Shared/graphql/utils.ts
@@ -1,12 +1,11 @@
 import { GRAPHQL_ENDPOINTS, TEndpointNames } from 'assets/env'
-import useSWR from 'swr'
 import { TypedDocumentString } from './index'
 
-async function fetcher<Data, Variables>([query, variables, endpointName]: [
-    TypedDocumentString<Data, Variables>,
-    Variables,
-    TEndpointNames,
-]) {
+export async function fetcher<Data, Variables>([
+    query,
+    variables,
+    endpointName,
+]: [TypedDocumentString<Data, Variables>, Variables, TEndpointNames]) {
     return fetch(GRAPHQL_ENDPOINTS[endpointName], {
         headers: {
             'Content-Type': 'application/json',
@@ -37,33 +36,4 @@ export async function fetchQuery<Data, Variables>(
         ...options,
     }
     return fetcher([query, variables, mergedOptions.endpoint])
-}
-
-type TUseQueryOptions = {
-    poll: boolean
-    endpoint: TEndpointNames
-}
-
-export function useQuery<Data, Variables>(
-    query: TypedDocumentString<Data, Variables>,
-    variables: Variables,
-    options?: Partial<TUseQueryOptions>,
-) {
-    const mergedOptions: TUseQueryOptions = {
-        poll: false,
-        endpoint: 'journey-planner',
-        ...options,
-    }
-
-    const { data, error, isLoading } = useSWR<Data>(
-        [query, variables, mergedOptions.endpoint],
-        fetcher,
-        {
-            revalidateOnFocus: false,
-            revalidateOnReconnect: true,
-            refreshInterval: mergedOptions.poll ? 30000 : undefined,
-        },
-    )
-
-    return { data, error, isLoading }
 }
